refactor(services): remove re-export shims and complete Phase 3-cleanup (WIP)
Delete 4 shim files and update all imports to use actual locations:
- abandon_flow_service.py → use flow.abandon directly
- flow_rebuild_usecase.py → use flow.rebuild directly
- handoff_service.py → use handoff.service directly
- convention_resolver.py → use vibe3.config directly

Updated imports in:
- orchestra/orchestrator.py
- flow/cleanup.py, flow/rebuild.py
- handoff/status.py
- orchestra/status.py
- pr/resolver.py
- issue/branch_resolver.py, issue/flow.py

All imports now use actual module paths instead of shim re-exports.
