--- src/vibe3/services/abandon_flow_service.py
@@ -1,5 +1,0 @@
-"""Re-export shim — AbandonFlowService has moved to vibe3.services.flow.abandon."""
-
-from vibe3.services.flow.abandon import AbandonFlowService
-
-__all__ = ["AbandonFlowService"]
