--- src/vibe3/services/convention_resolver.py
@@ -1,3 +1,0 @@
-from vibe3.config import ConventionResolver, get_convention, get_resolver
-
-__all__ = ["ConventionResolver", "get_convention", "get_resolver"]
