--- src/vibe3/services/flow_rebuild_usecase.py
@@ -1,5 +1,0 @@
-"""Re-export shim — FlowRebuildUsecase has moved to vibe3.services.flow.rebuild."""
-
-from vibe3.services.flow.rebuild import FlowRebuildUsecase
-
-__all__ = ["FlowRebuildUsecase"]
