+++ src/vibe3/services/handoff/status.py
@@ -8,7 +8,7 @@ from vibe3.clients import BackendProtocol, GitClient, SQLiteClient
 from vibe3.environment import SessionRegistryService
 from vibe3.models import FlowEvent, FlowState, VerdictRecord
 from vibe3.services.flow.service import FlowService
-from vibe3.services.handoff_service import HandoffService
+from vibe3.services.handoff.service import HandoffService
 from vibe3.services.pr.verdict_service import VerdictService
