--- src/vibe3/services/handoff_service.py
@@ -1,5 +1,0 @@
-"""Re-export shim — HandoffService has moved to vibe3.services.handoff.service."""
-
-from vibe3.services.handoff.service import HandoffService
-
-__all__ = ["HandoffService"]
