+++ src/vibe3/services/pr/resolver.py
@@ -5,8 +5,8 @@ import shutil
 import typer
 
 from vibe3.clients import GitHubClient
+from vibe3.config import get_convention
 from vibe3.exceptions import UserError
-from vibe3.services.convention_resolver import get_convention
 from vibe3.services.flow.service import FlowService
 from vibe3.services.issue.branch_resolver import resolve_issue_branch_input
